backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
supabase==2.0.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0